Added Update Wallet Functionality
Amended User Entity and added region to keep record of order's region.
Added:
- update wallet on Buy
- update wallet on Sell

+++ SpringSmartOrderRouting/src/main/java/com/ab/services/OrderService.java
@@ -61,7 +61,7 @@ public class OrderService {
 		if (!numeric) {
 			stockAmount = 0.0;
 		} else {
-			stockAmount = (double) Integer.parseInt(getNumber);
+			stockAmount = Double.parseDouble(getNumber);
 		}
 		return stockAmount;
 	}
